Remove everything to do with ExtensionPointSpec
Context here; I think previously the intent was that an Extension would define it's own extension points.  Because an extension can be inserted by a plugin, this meant that when defining an extension by a plugin, you would also define the extension points for that extension.  BUT! . Now that the logic provided by ExtensionPointSpec is infered, rather than being hard coded I don't think we need the ExtensionPointSpec plugin functionality anymore.

This does raise the question; What happens in future if an Extension wants to change the logic by which the surface is infered.  If that happens, I think we could move some logic to the Extension speicfication, rather than the additional overhead of the ExtensionSpec